src/test/java/com/study/testcode/programmers/IntegerTriangleTest.java
package com.study.testcode.programmers;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.TestPropertySource;

import static com.study.testcode.common.constant.DateConstants.CONFIG_TEST;
import static org.junit.jupiter.api.Assertions.assertEquals;

@SpringBootTest
@TestPropertySource(locations = CONFIG_TEST)
public class IntegerTriangleTest {
    @Autowired
    private IntegerTriangle integerTriangle;

    @Test
    void integerTriangleMaxValueTest() {
        int[][] triangle = {
                {7},
                {3, 8},
                {8, 1, 0},
                {2, 7, 4, 4},
                {4, 5, 2, 6, 5}
        };

        assertEquals(30, integerTriangle.mySolution(triangle));
        assertEquals(30, integerTriangle.bestSolution(triangle));
    }
}
